mysite/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
django-taggit==0.22.2
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
Markdown==2.6.11
Pillow==7.1.1
psycopg2==2.7.4
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.23.0
urllib3==1.25.9
whitenoise==5.0.1
